Add trip completed :D

+++ app/src/main/java/com/jets/mytrips/beans/Trip.java
@@ -22,7 +22,7 @@ public class Trip implements Parcelable{
     double endX;
     double endY;
     String date;
-    int time;
+    long time;
     String status;
     int done;
     ArrayList<Note> notes;
@@ -30,7 +30,7 @@ public class Trip implements Parcelable{
     int alarmId;
 
     public Trip(String id, int userId, String start, double startX, double startY, String end,
-                double endX, double endY, String date, int time, String status, int done,
+                double endX, double endY, String date, long time, String status, int done,
                 ArrayList<Note> notes, String image, int alarmId) {
         this.id = id;
         this.userId = userId;
@@ -64,7 +64,7 @@ public class Trip implements Parcelable{
         endX = in.readDouble();
         endY = in.readDouble();
         date = in.readString();
-        time = in.readInt();
+        time = in.readLong();
         status = in.readString();
         done = in.readInt();
         image = in.readString();
@@ -163,11 +163,11 @@ public class Trip implements Parcelable{
         this.date = date;
     }
 
-    public int getTime() {
+    public long getTime() {
         return time;
     }
 
-    public void setTime(int time) {
+    public void setTime(long time) {
         this.time = time;
     }
 
@@ -228,7 +228,7 @@ public class Trip implements Parcelable{
         dest.writeDouble(endX);
         dest.writeDouble(endY);
         dest.writeString(date);
-        dest.writeInt(time);
+        dest.writeLong(time);
         dest.writeString(status);
         dest.writeInt(done);
         dest.writeString(image);
